differential_dogs3::operators

Module half_join2

Functions

half_join

A binary equijoin that responds to updates on only its first input.

half_join_internal_unsafe

An unsafe variant of half_join where the output_func closure takes additional arguments a vector of time and diff tuples as input and writes its outputs at a container builder. The container builder can, but isn’t required to, accept (data, time, diff) triplets. This allows for more flexibility, but is more error-prone.