differential_dogs3::operators::lookup_map

Function lookup_map

pub fn lookup_map<'scope, D, K, R, Tr, F, DOut, ROut, S>(
    prefixes: VecCollection<'scope, Tr::Time, D, R>,
    arrangement: Arranged<'scope, Tr>,
    key_selector: F,
    output_func: S,
    supplied_key0: K,
    supplied_key1: K,
    supplied_key2: K,
) -> VecCollection<'scope, Tr::Time, DOut, ROut>
where
    Tr: for<'a> TraceReader<Time: Hash, Diff: Semigroup<Tr::DiffGat<'a>> + Monoid + ExchangeData> + Clone + 'static,
    Tr::KeyContainer: BatchContainer<Owned = K>,
    K: Hashable + Ord + 'static,
    F: FnMut(&D, &mut K) + Clone + 'static,
    D: ExchangeData,
    R: ExchangeData + Monoid,
    DOut: Clone + 'static,
    ROut: Monoid + 'static,
    S: FnMut(&D, &R, Tr::Val<'_>, &Tr::Diff) -> (DOut, ROut) + 'static,

Proposes extensions to a stream of prefixes.

This method takes a stream of prefixes and for each determines a key with key_selector and then proposes all pair af the prefix and values associated with the key in arrangement.