differential_dogs3::operators::propose

Function propose_distinct

pub fn propose_distinct<'scope, Tr, K, F, P, V>(
    prefixes: VecCollection<'scope, Tr::Time, P, Tr::Diff>,
    arrangement: Arranged<'scope, Tr>,
    key_selector: F,
) -> VecCollection<'scope, Tr::Time, (P, V), Tr::Diff>
where
    Tr: for<'a> TraceReader<ValOwn = V, Time: Hash, Diff: Semigroup<Tr::DiffGat<'a>> + Monoid + Multiply<Output = Tr::Diff> + ExchangeData> + Clone + 'static,
    Tr::KeyContainer: BatchContainer<Owned = K>,
    K: Hashable + Default + Ord + 'static,
    F: Fn(&P) -> K + Clone + 'static,
    P: ExchangeData,
    V: Clone + 'static,

Proposes distinct extensions to a prefix stream.

Unlike propose, this method does not scale the multiplicity of matched prefixes by the number of matches in arrangement. This can be useful to avoid the need to prepare an arrangement of distinct extensions.