differential_dogs3::operators::validate

Function validate

pub fn validate<'scope, K, V, Tr, F, P>(
    extensions: VecCollection<'scope, Tr::Time, (P, V), Tr::Diff>,
    arrangement: Arranged<'scope, Tr>,
    key_selector: F,
) -> VecCollection<'scope, Tr::Time, (P, V), Tr::Diff>
where
    Tr: for<'a> TraceReader<Time: Hash, Diff: Semigroup<Tr::DiffGat<'a>> + Monoid + Multiply<Output = Tr::Diff> + ExchangeData> + Clone + 'static,
    Tr::KeyContainer: BatchContainer<Owned = (K, V)>,
    K: Ord + Hash + Clone + Default + 'static,
    V: ExchangeData + Hash + Default,
    F: Fn(&P) -> K + Clone + 'static,
    P: ExchangeData,

Proposes extensions to a stream of prefixes.

This method takes a stream of prefixes and for each determines a key with key_selector and then proposes all pair af the prefix and values associated with the key in arrangement.