differential_dogs3

Trait PrefixExtender

pub trait PrefixExtender<'scope, T: Timestamp, R: Monoid + Multiply<Output = R>> {
    type Prefix;
    type Extension;

    // Required methods
    fn count(
        &mut self,
        prefixes: VecCollection<'scope, T, (Self::Prefix, usize, usize), R>,
        index: usize,
    ) -> VecCollection<'scope, T, (Self::Prefix, usize, usize), R>;
    fn propose(
        &mut self,
        prefixes: VecCollection<'scope, T, Self::Prefix, R>,
    ) -> VecCollection<'scope, T, (Self::Prefix, Self::Extension), R>;
    fn validate(
        &mut self,
        extensions: VecCollection<'scope, T, (Self::Prefix, Self::Extension), R>,
    ) -> VecCollection<'scope, T, (Self::Prefix, Self::Extension), R>;
}

A type capable of extending a stream of prefixes.

Implementors of PrefixExtension provide types and methods for extending a differential dataflow collection, via the three methods count, propose, and validate.

Required Associated Types

type Prefix

The required type of prefix to extend.

type Extension

The type to be produced as extension.

Required Methods

fn count( &mut self, prefixes: VecCollection<'scope, T, (Self::Prefix, usize, usize), R>, index: usize, ) -> VecCollection<'scope, T, (Self::Prefix, usize, usize), R>

Annotates prefixes with the number of extensions the relation would propose.

fn propose( &mut self, prefixes: VecCollection<'scope, T, Self::Prefix, R>, ) -> VecCollection<'scope, T, (Self::Prefix, Self::Extension), R>

Extends each prefix with corresponding extensions.

fn validate( &mut self, extensions: VecCollection<'scope, T, (Self::Prefix, Self::Extension), R>, ) -> VecCollection<'scope, T, (Self::Prefix, Self::Extension), R>

Restricts proposed extensions by those the extender would have proposed.